requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.1
Pillow==10.1.0
pydantic==2.5.2
pydantic-core==2.14.5
python-dotenv==1.0.0
sqlparse==0.4.4
typing-extensions==4.8.0
